requirements.txt
flask==3.0.0
python-dotenv==1.0.1
langchain==0.1.12
langchain-openai==0.0.8
requests==2.31.0
uvicorn==0.27.1
asgiref==3.7.2
pydantic==2.6.3
openai==1.13.3
fastapi==0.110.0
python-multipart==0.0.9 